+++ tests/duration_tests/test_total_methods.py
@@ -1,4 +1,4 @@
-from pendulum.interval import Interval
+from pendulum.duration import Duration
 
 from .. import AbstractTestCase
 
@@ -6,21 +6,21 @@ from .. import AbstractTestCase
 class TotalMethodsTest(AbstractTestCase):
 
     def test_in_weeks(self):
-        it = Interval(days=17)
+        it = Duration(days=17)
         self.assertEqual(2.43, round(it.total_weeks(), 2))
 
     def test_in_days(self):
-        it = Interval(days=3)
+        it = Duration(days=3)
         self.assertEqual(3, it.total_days())
 
     def test_in_hours(self):
-        it = Interval(days=3, minutes=72)
+        it = Duration(days=3, minutes=72)
         self.assertEqual(73.2, it.total_hours())
 
     def test_in_minutes(self):
-        it = Interval(minutes=6, seconds=72)
+        it = Duration(minutes=6, seconds=72)
         self.assertEqual(7.2, it.total_minutes())
 
     def test_in_seconds(self):
-        it = Interval(seconds=72, microseconds=123456)
+        it = Duration(seconds=72, microseconds=123456)
         self.assertEqual(72.123456, it.total_seconds())